% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bvr_stations}
\alias{bvr_stations}
\title{BVR Water Quality Stations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33 rows and 7 columns.}
\usage{
bvr_stations
}
\description{
Water Quality Stations for Big Valley Rancheria
\describe{
  \item{origin_id}{id for the orgnization collecting/supplying data}
  \item{station_id}{id for the station}
  \item{lat}{latitude}
  \item{lon}{longitude}
  \item{station_horizontal_datum}{horizontal datum for the station}
  \item{state}{state}
  \item{county}{county}
}
}
\keyword{datasets}
