% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bvr_water_quality}
\alias{bvr_water_quality}
\title{BVR Water Quality Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19894 rows and 10 columns.}
\usage{
bvr_water_quality
}
\description{
Water Quality Data measured by Big Valley Rancheria (BVR). This data
contains only analytes that were measured at least 10 times.
\describe{
  \item{origin_id}{id for the orgnization collecting/supplying data}
  \item{station_id}{id for the station (assume unique only within this orgin_id)}
  \item{datetime}{the date and time of observation}
  \item{analyte}{name of the analyte measured}
  \item{units}{units reported for the measurement}
  \item{sample_fraction}{blah}
  \item{value_type}{a qualifier for the type of value}
  \item{statistic_type}{summary statistic used in result, if none this is NA}
  \item{value_raw}{raw value reported from the measurement}
  \item{value_numeric}{the numerical value obtained from value raw}
}
}
\keyword{datasets}
