% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bvr_wq}
\alias{bvr_wq}
\title{BVR Water Quality}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19614 rows and 12 columns.}
\usage{
bvr_wq
}
\description{
BVR Water Quality
}
\keyword{datasets}
