% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cache_creek_flow}
\alias{cache_creek_flow}
\title{Cache Creek near Lower Lake Flow}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7155 rows and 4 columns.}
\usage{
cache_creek_flow
}
\description{
Flow at Cache Creek near Lower Lake using site no: 11451000 from USGS. Data is
reported as the daily average.
\describe{
  \item{site_no}{site number from USGS}
  \item{site_name}{name for the site}
  \item{date}{date for observed flow}
  \item{flow_cfs}{flow value in cfs}
}
}
\keyword{datasets}
