% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfa_stations}
\alias{cdfa_stations}
\title{CDFA Water Quality Stations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2083 rows and 4 columns.}
\usage{
cdfa_stations
}
\description{
Water Quality Stations for the California Department of Food and Ag (CDFA)
\describe{
  \item{origin_id}{id for the data supplier}
  \item{station_id}{id for the sttation (assume unique only within origin id)}
  \item{lat}{station latitude}
  \item{lon}{station longitude}
}
}
\keyword{datasets}
