% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdfa_water_quality}
\alias{cdfa_water_quality}
\title{CDFA Water Quality Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2083 rows and 7 columns.}
\usage{
cdfa_water_quality
}
\description{
Water Quality data reported by the CDFA
\describe{
  \item{origin_id}{id for the data supplier}
  \item{station_id}{id for observation station}
  \item{datetime}{datetime of measurement}
  \item{unit}{unit of measure}
  \item{value_raw}{raw value obtained from the measurement}
  \item{value_numeric}{numeric version of the measurement}
  \item{method_name}{method used for the measurement}
}
}
\keyword{datasets}
