% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ceden_stations}
\alias{ceden_stations}
\title{CEDEN Water Quality Stations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3143 rows and 4 columns.}
\usage{
ceden_stations
}
\description{
fda
\describe{
  \item{origin_id}{id for the data supplier}
  \item{station_id}{id for the station}
  \item{lat}{latitude}
  \item{lon}{longitude}
}
}
\keyword{datasets}
