% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ceden_water_quality}
\alias{ceden_water_quality}
\title{CEDEN Water Quality Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3143 rows and 8 columns.}
\usage{
ceden_water_quality
}
\description{
water quality data reported by Aquatic Pesticide Monitoring Program,
Irrigated Lands Regulatory Program and Surface Water Ambient Monitoring Program
obtained through CEDEN.
\describe{
  \item{origin_id}{program or orgnization collecting data}
  \item{station_id}{id for station where data was measured}
  \item{datetime}{datetime of measurement}
  \item{analyte}{name of the analyte measured}
  \item{unit}{unit reported in the measurement}
  \item{value_numeric}{numeric value obtained from measurement}
  \item{matrix_name}{blah}
  \item{result_qual_code}{blah}
}
}
\keyword{datasets}
