% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clear_lake_wse}
\alias{clear_lake_wse}
\title{Clear Lake Water Surface Elevation}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7513 rows and 4 columns.}
\usage{
clear_lake_wse
}
\description{
Clear Lake Water Surface Elevation
}
\keyword{datasets}
