% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_stations}
\alias{flow_stations}
\title{Flow Stations}
\format{An object of class \code{data.frame} with 5 rows and 5 columns.}
\usage{
flow_stations
}
\description{
Flow stations around Big Valley Rancheria
\describe{
  \item{agency_cd}{agency code}
  \item{site_no}{site number from USGS}
  \item{site_name}{name for the site}
  \item{lat}{latitide}
  \item{lon}{longitude}
}
}
\keyword{datasets}
