% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groundwater_levels}
\alias{groundwater_levels}
\title{Groundwater Elevations}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7154 rows and 7 columns.}
\usage{
groundwater_levels
}
\description{
These are groundwater elevations for wells located in the
Big Valley Water Basin. The data was obtained from CASGEM.
\describe{
  \item{site_code}{a unique CASGEM code for the well}
  \item{measurement_date}{date when measure was taken}
  \item{rp_elevation}{reference point of elevation}
  \item{gs_elevation}{groun surface elevation}
  \item{ws_reading}{water surface reading}
  \item{rp_reading}{reference point reading}
}
}
\keyword{datasets}
