% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groundwater_stations}
\alias{groundwater_stations}
\title{Groundwater Stations}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 85 rows and 8 columns.}
\usage{
groundwater_stations
}
\description{
These are the wells for which data is reported/
\describe{
  \item{site_code}{a unique CASGEM code for the well}
  \item{local_well_designation}{well designation for local use}
  \item{lat}{latitude for the well}
  \item{lng}{longitude for the well}
  \item{total_depth_ft}{total depth for the well (if not confidential)}
  \item{start_date}{first date measured (with CASGEM data)}
  \item{end_date}{most recent data with measured data}
  \item{total_measures}{number of measurements taken in date range}
}
}
\keyword{datasets}
