% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kelsey_creek_flow}
\alias{kelsey_creek_flow}
\title{Kelsey Creek Flow}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18113 rows and 4 columns.}
\usage{
kelsey_creek_flow
}
\description{
Flow at Kelsey Creek using site no: 11449500 from USGS. Data is
reported as the daily average.
\describe{
  \item{site_no}{site number from USGS}
  \item{site_name}{name for the site}
  \item{date}{date for observed flow}
  \item{flow_cfs}{flow value in cfs}
}
}
\keyword{datasets}
