% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pressure_transducer}
\alias{pressure_transducer}
\title{Pressure Transducer Data}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 29101 rows and 4 columns.}
\usage{
pressure_transducer
}
\description{
blah
\describe{
  \item{transducer}{name of the transducer}
  \item{dateTime}{date-time for the measurements}
  \item{depth_ft}{the depth (in feet) reported by the pressure transducer}
  \item{temp_c}{the temperature (in C) reported by the pressure transducer}
}
}
\keyword{datasets}
