% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bvr_stations}
\alias{bvr_stations}
\title{BVR Water Quality Stations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33 rows and 8 columns.}
\usage{
bvr_stations
}
\description{
BVR Water Quality Stations
}
\keyword{datasets}
