% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hitch_counts}
\alias{hitch_counts}
\title{Hitch Counts}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 773 rows and 6 columns.}
\usage{
hitch_counts
}
\description{
longer
\describe{
  \item{origin_id}{program or orgnization collecting data}
}
}
\details{
details abpout the data
}
\keyword{datasets}
